% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absolute.R
\name{absolute}
\alias{absolute}
\alias{absolute.list}
\alias{absolute.deafult}
\title{Absolute}
\usage{
absolute(x)

\method{absolute}{list}(x)

\method{absolute}{deafult}(x)
}
\arguments{
\item{x}{fainress object}
}
\value{
summed absolute difference betwen ideal measure (1) per group
}
\description{
Absolute
}
