% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_adult_test.R
\docType{data}
\name{adult_test}
\alias{adult_test}
\title{Adult test dataset}
\format{
A data frame with 16281 rows and 15 variables:
\describe{
  \item{salary}{factor, <=50K/>50K whether a person salary exceeds 50K a year or not}
  \item{age}{integer, age of person}
  \item{workclass}{factor, field of work}
  \item{fnlwgt}{numeric}
  \item{education}{factor, completed education degree}
  \item{education_num}{numeric, education number in converted from education factor, the bigger the better}
  \item{marital_status}{factor}
  \item{occupation}{factor, where this person works}
  \item{relationship}{factor, relationship information}
  \item{race}{factor, ethnicity of a person}
  \item{sex}{factor, gender of a person}
  \item{capital_gain}{numeric}
  \item{capital_loss}{numeric}
  \item{hours_per_week}{numeric, how many hours per week does this person work}
  \item{native_country}{factor, in which country was this person born}
}
}
\source{
Data from UCL \url{https://archive.ics.uci.edu/ml/datasets/adult}
}
\usage{
data(adult_test)
}
\description{
\code{adult_test}
}
