% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_cutoffs.R
\name{all_cutoffs}
\alias{all_cutoffs}
\title{All cutoffs}
\usage{
all_cutoffs(x, label, grid_points = 101, fairness_metrics = unique_metrics())
}
\arguments{
\item{x}{fairness_object}

\item{label}{label of model to be calculated}

\item{grid_points}{numeric, grid for cutoffs to test. Number of points between 0 and 1 spread evenly.}

\item{fairness_metrics}{character, name of metric or vector of multiple metrics}
}
\value{
all_cutoffs object
}
\description{
Create all_cutoffs object and to show how with change of cutoffs parity metrics change for single model/explainer.
It is highly suggested to pick small number of metrics. With value of cutoff all subgroup cutoffs change.
}
\examples{

data("compas")

# positive outcome - not being recidivist
two_yr_recidivism <- factor(compas$Two_yr_Recidivism, levels = c(1,0))
y_numeric <- as.numeric(two_yr_recidivism) -1

lm_model <- glm(Two_yr_Recidivism~.,
                data=compas,
                family=binomial(link="logit"))

rf_model <- ranger::ranger(Two_yr_Recidivism ~.,
                           data = compas,
                           probability = TRUE,
                           num.trees = 200)

explainer_lm <- DALEX::explain(lm_model, data = compas[,-1], y = y_numeric)
explainer_rf <- DALEX::explain(rf_model, data = compas[,-1], y = y_numeric)

fobject <- fairness_check(explainer_lm, explainer_rf,
                          protected = compas$Ethnicity,
                          privileged = "Caucasian")

ac <- all_cutoffs(fobject,
                  label = "lm",
                  fairness_metrics = c("TPR_parity_loss",
                                       "FPR_parity_loss"))
plot(ac)

}
