% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_group_fairness_metrics.R,
%   R/create_group_metric_matrix.R
\name{calculate_group_fairness_metrics}
\alias{calculate_group_fairness_metrics}
\title{Calculate fairness metrics in groups}
\usage{
calculate_group_fairness_metrics(x)

calculate_group_fairness_metrics(x)
}
\arguments{
\item{x}{object of class \code{group_matrices}}
}
\value{



}
\description{
Calculate fairness metrics in groups

Calculate fairness metrics in groups
}
