% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_metric.R
\name{choose_metric}
\alias{choose_metric}
\title{Choose metric}
\usage{
choose_metric(x, fairness_metric = "FPR_parity_loss")
}
\arguments{
\item{x}{fairness object}

\item{fairness_metric}{\code{char}, name of fairness metric, one of metric data in fairness object}
}
\value{
choose_metric object
}
\description{
Choose metric
}
\examples{

library(DALEX)
library(ranger)

data(compas)

rf_compas <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE) # Wszystko
lr_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))
# numeric target values
y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

# explainer
rf_explainer <- explain(rf_compas, data = compas[,-1], y = y_numeric)
lr_explainer <- explain(lr_compas, data = compas[,-1], y = y_numeric)


fobject <- create_fairness_object(rf_explainer, lr_explainer,
                                  data = compas,
                                  outcome = "Two_yr_Recidivism",
                                  group = "Ethnicity",
                                  base = "African_American")

choose_metric(fobject, "TPR_parity_loss")
}
