% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_parity.R
\name{class_parity}
\alias{class_parity}
\alias{class_parity.fairness_object}
\alias{class_parity.default}
\title{Class parity}
\usage{
class_parity(x, ...)

\method{class_parity}{fairness_object}(x, ...)

\method{class_parity}{default}(x, ..., outcome, group)
}
\arguments{
\item{x}{data.frame/ fairness_object}

\item{...}{other parameters}

\item{outcome}{target column name in data}

\item{group}{protected group (column) in data}
}
\value{

}
\description{
Class parity can be thought as demographic parity for data. It does not look at predictions but original labels in data. It can visualize parity on dataset level. In perfect dataset classes should be evenly distributed across subgroups. Plot allows
user to see not only how classes are distributed but also the size of subgroups. Plot allows to localize problems in dataset level. Function can receive data frame or fairness object.
}
\examples{

data("compas")
class_parity(compas, outcome =  "Two_yr_Recidivism", group = "Ethnicity")

}
