% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_group_metric_matrix.R
\name{create_group_metric_matrix}
\alias{create_group_metric_matrix}
\title{Metrics for groups}
\usage{
create_group_metric_matrix(x)
}
\arguments{
\item{x}{object of class \code{group_matrices}}
}
\value{

}
\description{
Metrics for groups
}
