% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_parity.R
\name{demographic_parity}
\alias{demographic_parity}
\alias{demographic_parity.fairness_object}
\alias{demographic_parity.default}
\title{Demographic parity}
\usage{
demographic_parity(x, ...)

\method{demographic_parity}{fairness_object}(x, ...)

\method{demographic_parity}{default}(x, ..., outcome, group)
}
\arguments{
\item{x}{data.frame/ fairness_object}

\item{...}{other parameters}

\item{outcome}{target column name in data}

\item{group}{protected group (column) in data}
}
\value{

}
\description{
Demographic parity is popular choice to visualize parity on dataset level. In perfect dataset classes should be evenly distributed across subgroups. Plot allows
user to see not only how classes are distributed but also the size of subgroups. Plot allows to localize problems in dataset level. Function can recieve data frame or fairness object.
}
\examples{

data("compas")
demographic_parity(compas, outcome =  "Two_yr_Recidivism", group = "Ethnicity")

}
