% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairness_check.R
\name{fairness_check}
\alias{fairness_check}
\title{Fairness check}
\usage{
fairness_check(x, epsilon = 0.1)
}
\arguments{
\item{x}{fairness_object}

\item{epsilon}{numeric, margins for decision if fair or not. Default 0.1}
}
\value{

}
\description{
Method for quick popular fairness metrics check. On contrary to parity loss metrics here fairness_check bases on group_metrics
and worse performance in particular metric means, that subgroup scored worse than privileged (base) subgroup. Note that being worse in particular
metric depends on assumption that models try to predict positive outcome. Therefore Equal odds here calculates True positive rate. Unlike in Parity
loss metrics assigning subjectively negative metric as positive (numeric 1) will end up in method's wrong interpretation of outcome. When plotted
shows 5 metric scores for each subgroups and each model.
}
\details{
Metrics used are made for each subgroup, then base metric score is subtracted leaving loss of particular metric.
If loss is less than -epsilon than such metric is marked as "not passed". There is no upper boundary for fairness, so if some subgroup is "better"
in particular metric than privileged one there are no consequences to that. Good metric description can be found here :
\url{https://en.wikipedia.org/wiki/Fairness_(machine_learning)}
}
\examples{

library(DALEX)
library(ranger)

data("compas")

# chaging levels - positive situation here is not becoming recidivist
compas$Two_yr_Recidivism <- factor(compas$Two_yr_Recidivism, levels = c(1,0))

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fobject <-create_fairness_object(explainer_glm, explainer_rf,
                                 outcome = "Two_yr_Recidivism",
                                 data  = compas,
                                 group = "Sex",
                                 base  = "Female",
                                cutoff = 0.5)

fc <- fairness_check(fobject)
plot(fc)

}
\references{
Zafar,Valera, Rodriguez, Gummadi (2017)  \url{https://arxiv.org/pdf/1610.08452.pdf}

Hardt, Price, Srebro (2016) \url{https://arxiv.org/pdf/1610.02413.pdf}
}
