% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairness_check_regression.R
\name{fairness_check_regression}
\alias{fairness_check_regression}
\title{Fairness check regression}
\usage{
fairness_check_regression(
  x,
  ...,
  protected = NULL,
  privileged = NULL,
  label = NULL,
  epsilon = NULL,
  verbose = TRUE,
  colorize = TRUE
)
}
\arguments{
\item{colorize}{}
}
\value{

}
\description{
Fairness check regression
}
\examples{


}
