% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairness_radar.R
\name{fairness_radar}
\alias{fairness_radar}
\title{Fairness radar}
\usage{
fairness_radar(x, metrics = NULL)
}
\arguments{
\item{x}{\code{fairness_object}}

\item{metrics}{vector of metric names, at least 3 metrics without NA needed. If \code{NULL} deafult metrics will be taken}
}
\value{
\code{fairness_radar} object
}
\description{
Make fairness_radar object whith chosen metrics. Note that there must be at least three metrics that does not contain NA.
}
\examples{
library(DALEX)
library(ranger)

data("compas")

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fobject <- create_fairness_object(explainer_glm, explainer_rf,
                             outcome = "Two_yr_Recidivism",
                             group = "Ethnicity",
                             base = "Caucasian",
                             cutoff = 0.5)
fradar <- fairness_radar(fobject)
plot(fradar)
}
