% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_objects.R
\name{get_objects}
\alias{get_objects}
\title{Get objects}
\usage{
get_objects(x, class)
}
\arguments{
\item{x}{list of elements}

\item{class}{character, name of class to be extracted from list}
}
\value{

}
\description{
Extract certain class objects from a list
}
