% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_metric.R
\name{group_metric}
\alias{group_metric}
\title{Group metric}
\usage{
group_metric(
  x,
  fairness_metric = NULL,
  performance_metric = NULL,
  parity_loss = FALSE
)
}
\arguments{
\item{x}{object of class \code{fairness_object}}

\item{fairness_metric}{character, fairness metric name}

\item{performance_metric}{character, performance metric name}

\item{parity_loss}{logical, if TRUE parity loss will supersede normal metric}
}
\value{
\code{group_metric} object
}
\description{
Group metric enables to extract data from metrics generated for each group, and prepare it for visualization.
}
\details{
Where:

Fairness metrics:

\itemize{
\item TPR
\item TNR
\item PPV
\item NPV
\item FNR
\item FPR
\item FDR
\item FOR
\item TS
\item ACC
\item F1
\item MCC
}
Performance metrics

\itemize{
\item recall
\item precision
\item accuracy
\item f1
\item auc
}
}
\examples{

library(DALEX)
library(ranger)
data("compas")
y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

rf_1 <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
lr_1 <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

explainer_1 <- explain(rf_1, data = compas, y = y_numeric)
explainer_2 <- explain(lr_1, data = compas, y = y_numeric)

fo <- create_fairness_object(explainer_1, explainer_2,  outcome = "Two_yr_Recidivism", group = "Ethnicity", base = "Caucasian"  )

gm <- group_metric(fo, fairness_metric = "FPR", performance_metric = "f1")
plot(gm)

}
