% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmap}
\alias{heatmap}
\alias{heatmap.fairness_object}
\alias{heatmap.deafult}
\title{Heatmap}
\usage{
heatmap(x, midpoint = NULL, ...)

\method{heatmap}{fairness_object}(x, midpoint = NULL, ...)

\method{heatmap}{deafult}(x, midpoint = NULL, ...)
}
\arguments{
\item{x}{fairness object}

\item{midpoint}{point in score scale}

\item{...}{other parameters}
}
\value{
ggplot object
}
\description{
Heatmap
}
\examples{
library(ranger)
library(DALEX)
df <- fairness::compas[,1:7]
# models
y_numeric <- as.numeric(df$Two_yr_Recidivism)-1

rf_compas_1 <- ranger(Two_yr_Recidivism ~Number_of_Priors+Age_Below_TwentyFive, data = df, probability = TRUE) # Wszystko
model_compas_lr <- glm(Two_yr_Recidivism~., data=df, family=binomial(link="logit"))
rf_compas_5 <- ranger(Two_yr_Recidivism ~., data = df, probability = TRUE) # Wszystko
rf_compas_6 <- ranger(Two_yr_Recidivism ~ Age_Above_FourtyFive+Misdemeanor, data = df, probability = TRUE) # dziwny model
rf_compas_7 <- ranger(Two_yr_Recidivism ~., data = df, probability = TRUE) # Wszystko
rf_compas_8 <- ranger(Two_yr_Recidivism ~ Female+Age_Above_FourtyFive+Misdemeanor+ethnicity, data = df, probability = TRUE) # dziwny model

# explainers
explainer_1 <- explain(rf_compas_1, data = df, y = y_numeric)
explainer_4 <- explain(model_compas_lr, data = df, y = y_numeric)
explainer_5 <- explain(rf_compas_5, data = df, y = y_numeric)
explainer_6 <- explain(rf_compas_6, data = df, y = y_numeric)
explainer_7 <- explain(rf_compas_7, data = df, y = y_numeric)

# diffrent labels
explainer_4$label <- "glm"
explainer_5$label <- "rf5"
explainer_6$label <- "rf6"
explainer_7$label <- "rf7"


fobject <- create_fairness_object(explainer_1,explainer_4,explainer_5,explainer_6,explainer_7,
                                  outcome = "Two_yr_Recidivism",
                                  group  = "ethnicity",
                                  base   = "Caucasian")

heatmap(fobject)
}
