% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric_scores.R
\name{metric_scores}
\alias{metric_scores}
\title{Metric scores}
\usage{
metric_scores(x, fairness_metrics = unique_metrics())
}
\arguments{
\item{x}{object of class \code{fairness_object}}

\item{fairness_metrics}{character, vector with fairness metric names}
}
\value{
\code{metric_scores} object.
It is a list containing:
\itemize{
\item{metric_scores_data}{ - \code{data.frame} with information about score in particular subgroup, metric, and model}
\item{privileged}{ - name of privileged subgroup}
}
}
\description{
Creates \code{metric_scores} object to facilitate visualization. Check how the metric scores differ among models, what is this score, and how it changes
for example after applying bias mitigation technique. The vertical black lines
denote the scores for privileged subgroup. It is best to use only few metrics (using \code{fairness_metrics} parameter)
}
\examples{

data("compas")

# flipping outcomes, 1- favorable
compas$Two_yr_Recidivism <- as.factor(ifelse(compas$Two_yr_Recidivism == '1', '0', '1'))

# train
lm_compas <- glm(Two_yr_Recidivism ~., data = compas, family = binomial())
rf_compas <- ranger::ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)

# numeric target values
y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

# explainer
rf_explainer <- DALEX::explain(rf_compas, data = compas[,-1], y = y_numeric)
lm_explainer <- DALEX::explain(lm_compas, data = compas[,-1], y = y_numeric)

fobject <- fairness_check(rf_explainer, lm_explainer,
                          protected = compas$Ethnicity,
                          privileged = "Caucasian")

ms <- metric_scores(fobject, fairness_metrics = c("TPR","STP","ACC"))
plot(ms)

}
