% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_with_fairness.R
\name{performance_and_fairness}
\alias{performance_and_fairness}
\title{Performance and fairness}
\usage{
performance_and_fairness(x, fairness_metric = NULL, performance_metric = NULL)
}
\arguments{
\item{x}{\code{fairness_object}}

\item{fairness_metric}{fairness metric, one of those in fairness_object}

\item{performance_metric}{performance metric, one of}
}
\value{
\code{performance_and_fairness} object
}
\description{
measure model performance and model fairness metric at the same time. Choose best model according to bot metrics.

Measure performance in both fairness metric and
}
\examples{

library(DALEX)
library(ranger)

data("compas")

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

rf_compas_1 <- ranger(Two_yr_Recidivism ~Number_of_Priors+Age_Below_TwentyFive, data = compas, probability = TRUE)
lr_compas_1 <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))
rf_compas_2 <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
rf_compas_3 <- ranger(Two_yr_Recidivism ~ Age_Above_FourtyFive+Misdemeanor, data = compas, probability = TRUE)
df <- compas
df$Two_yr_Recidivism <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_1 <- explain(rf_compas_1,  data = compas[,-1], y = y_numeric)
explainer_2 <- explain(lr_compas_1,  data = compas[,-1], y = y_numeric)
explainer_3 <- explain(rf_compas_2,  data = compas[,-1], y = y_numeric, label = "ranger_2")
explainer_4 <- explain(rf_compas_3,  data = compas[,-1], y = y_numeric, label = "ranger_3")

fobject <- create_fairness_object(explainer_1,explainer_2,explainer_3,explainer_4,
                                    data = compas,
                                    outcome = "Two_yr_Recidivism",
                                    group  = "Ethnicity",
                                    base   = "Caucasian")
paf <- performance_and_fairness(fobject)
plot(paf)

}
