% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fairness_check.R
\name{plot.fairness_check}
\alias{plot.fairness_check}
\title{Plot fairness check}
\usage{
\method{plot}{fairness_check}(x, ...)
}
\arguments{
\item{x}{fainress_check object}

\item{...}{other plot parameters}
}
\value{
ggplot object
}
\description{
Plot fairness check enables to look how big differences are between base subgroup (privildged) and unpriviledged ones.
If barplot reaches red zone it means that for this subgroup fairness goal is not satisfied. Multiple subgroups and models can be plotted.
Red and green zone boundary can be moved through epsilon parameter, that needs to be passed through fairness_check.
}
\examples{

library(DALEX)
library(ranger)

data("compas")

# chaging levels - positive situation here is not becoming recidivist
compas$Two_yr_Recidivism <- factor(compas$Two_yr_Recidivism, levels = c(1,0))

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fobject <-create_fairness_object(explainer_glm, explainer_rf,
                                 outcome = "Two_yr_Recidivism",
                                 data  = compas,
                                 group = "Sex",
                                 base  = "Female",
                                cutoff = 0.5)

fc <- fairness_check(fobject)
plot(fc)
}
