% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fairness_check.R
\name{plot.fairness_object}
\alias{plot.fairness_object}
\title{Plot fairness check}
\usage{
\method{plot}{fairness_object}(x, ...)
}
\arguments{
\item{x}{\code{fairness_check} object}

\item{...}{other \code{fairness_check} objects}
}
\value{
ggplot object
}
\description{
Plot fairness check enables to look how big differences are between base subgroup (privileged) and unprivileged ones.
If barplot reaches red zone it means that for this subgroup fairness goal is not satisfied. Multiple subgroups and models can be plotted.
Red and green zone boundary can be moved through epsilon parameter, that needs to be passed through \code{fairness_check}. Plot can receive many objects that
will be plotted together with condition of equal parameters (number of subgroups, epsilon, etc...).
}
\examples{

library(DALEX)
library(ranger)

data("compas")

# chaging levels - positive situation here is not becoming recidivist
compas$Two_yr_Recidivism <- factor(compas$Two_yr_Recidivism, levels = c(1,0))

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fobject <- fairness_check(explainer_glm, explainer_rf,
                                 protected  = compas$Sex,
                                 privileged = "Female")

plot(fobject)
}
