% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_group_metric.R
\name{plot.group_metric}
\alias{plot.group_metric}
\title{Plot group metric}
\usage{
\method{plot}{group_metric}(x, ...)
}
\arguments{
\item{x}{object of class group_metric}

\item{...}{other params}
}
\value{
list of \code{ggplot} object
}
\description{
Plot choosen metric in group. Notice how models are treatung different subgroups. Compare models both in fairness metrics and in performance.
}
\examples{

library(DALEX)
library(ranger)
data("compas")
y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

rf_1 <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
lr_1 <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

explainer_1 <- explain(rf_1, data = compas, y = y_numeric)
explainer_2 <- explain(lr_1, data = compas, y = y_numeric)

fo <- create_fairness_object(explainer_1, explainer_2,  outcome = "Two_yr_Recidivism", group = "Ethnicity", base = "Caucasian"  )

gm <- group_metric(fo, fairness_metric = "FPR", performance_metric = "auc")
plot(gm)

}
