% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_choosen_metric.R
\name{plot.choosen_metric}
\alias{plot.choosen_metric}
\title{Plot choosen metric}
\usage{
\method{plot}{choosen_metric}(x, ...)
}
\arguments{
\item{x}{object of class \code{choosen_metric}}

\item{...}{other plot parameters}
}
\value{
\code{ggplot2} object
}
\description{
Choose metric and than plot it for every model.
}
\examples{

library(DALEX)
library(ranger)

data(compas)

rf_compas <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE) # Wszystko
lr_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))
# numeric target values
y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

# explainer
rf_explainer <- explain(rf_compas, data = compas[,-1], y = y_numeric)
lr_explainer <- explain(lr_compas, data = compas[,-1], y = y_numeric)


fobject <- create_fairness_object(rf_explainer, lr_explainer,
                                  data = compas,
                                  outcome = "Two_yr_Recidivism",
                                  group = "Ethnicity",
                                  base = "African_American")

cm <- choose_metric(fobject, "TPR_parity_loss")
plot(cm)
}
