% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_chosen_metric.R
\name{plot.chosen_metric}
\alias{plot.chosen_metric}
\title{Plot chosen metric}
\usage{
\method{plot}{chosen_metric}(x, ...)
}
\arguments{
\item{x}{object of class \code{chosen_metric}}

\item{...}{other objects of class \code{chosen_metric}}
}
\value{
\code{ggplot2} object
}
\description{
Choose metric from parity loss metrics and plot it for every model.
 The one with the least parity loss is more fair in terms of this particular metric.
}
\examples{

data("compas")

# positive outcome - not being recidivist
two_yr_recidivism <- factor(compas$Two_yr_Recidivism, levels = c(1,0))
y_numeric <- as.numeric(two_yr_recidivism) -1

lm_model <- glm(Two_yr_Recidivism~.,
                data=compas,
                family=binomial(link="logit"))

rf_model <- ranger::ranger(Two_yr_Recidivism ~.,
                           data = compas,
                           probability = TRUE,
                           num.trees = 200)

explainer_lm <- DALEX::explain(lm_model, data = compas[,-1], y = y_numeric)
explainer_rf <- DALEX::explain(rf_model, data = compas[,-1], y = y_numeric)

fobject <- fairness_check(explainer_lm, explainer_rf,
                          protected = compas$Ethnicity,
                          privileged = "Caucasian")

cm <- choose_metric(fobject, "TPR_parity_loss")
plot(cm)
}
