% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_performance_and_fairness.R
\name{plot.performance_and_fairness}
\alias{plot.performance_and_fairness}
\title{Plot fairness and performance}
\usage{
\method{plot}{performance_and_fairness}(x, ...)
}
\arguments{
\item{x}{fairness object}

\item{...}{other plot parameters}
}
\value{

}
\description{
visualise fairness and model metric at the same time. Note that fairness metric parity scale is reversed so that the best models are in top right corner.
}
\examples{

library(DALEX)
library(ranger)

data("compas")

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

rf_compas_1 <- ranger(Two_yr_Recidivism ~Number_of_Priors+Age_Below_TwentyFive, data = compas, probability = TRUE)
lr_compas_1 <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))
rf_compas_2 <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
rf_compas_3 <- ranger(Two_yr_Recidivism ~ Age_Above_FourtyFive+Misdemeanor, data = compas, probability = TRUE)
df <- compas
df$Two_yr_Recidivism <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_1 <- explain(rf_compas_1,  data = compas[,-1], y = y_numeric)
explainer_2 <- explain(lr_compas_1,  data = compas[,-1], y = y_numeric)
explainer_3 <- explain(rf_compas_2,  data = compas[,-1], y = y_numeric, label = "ranger_2")
explainer_4 <- explain(rf_compas_3,  data = compas[,-1], y = y_numeric, label = "ranger_3")

fobject <- create_fairness_object(explainer_1,explainer_2,explainer_3,explainer_4,
                                    data = compas,
                                    outcome = "Two_yr_Recidivism",
                                    group  = "Ethnicity",
                                    base   = "Caucasian")


paf <- performance_and_fairness(fobject)
plot(paf)
}
