% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fairness_heatmap.R
\name{plot.fairness_heatmap}
\alias{plot.fairness_heatmap}
\title{Plot Heatmap}
\usage{
\method{plot}{fairness_heatmap}(
  x,
  ...,
  midpoint = NULL,
  title = NULL,
  subtitle = NULL,
  text = TRUE,
  scale = FALSE,
  flip_axis = FALSE
)
}
\arguments{
\item{x}{fairness object}

\item{...}{other plot parameters}

\item{midpoint}{midpoint on gradient scale}

\item{title}{title of the plot}

\item{subtitle}{subtitle of the plot}

\item{text}{deafult \code{TRUE} means it shows values on tiles}

\item{scale}{whether metrics should be normalised}

\item{flip_axis}{logical, whether to change axis with metrics on axis with models}
}
\value{
ggplot object
}
\description{
Heatmap shows all metrics across all models while displaying similarity between variables (in form of dendograms).
Heatmap is flexible and is always longer than wider for esthetic reasons. It flips axis while number of explainers exceeds
number of metrics. NA's in metrics are gray.
}
\examples{

library(ranger)
library(DALEX)
data(compas)

# models
y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

rf_compas_1 <- ranger(Two_yr_Recidivism ~Number_of_Priors+Age_Below_TwentyFive, data = compas, probability = TRUE)
model_compas_lr <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))
rf_compas_5 <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
rf_compas_6 <- ranger(Two_yr_Recidivism ~ Age_Above_FourtyFive+Misdemeanor, data = compas, probability = TRUE)
rf_compas_7 <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
rf_compas_8 <- ranger(Two_yr_Recidivism ~ Sex+Age_Above_FourtyFive+Misdemeanor+Ethnicity, data = compas, probability = TRUE)

# explainers
explainer_1 <- explain(rf_compas_1, data = compas, y = y_numeric)
explainer_4 <- explain(model_compas_lr, data = compas, y = y_numeric)
explainer_5 <- explain(rf_compas_5, data = compas, y = y_numeric)
explainer_6 <- explain(rf_compas_6, data = compas, y = y_numeric)
explainer_7 <- explain(rf_compas_7, data = compas, y = y_numeric)

# different labels
explainer_4$label <- "glm"
explainer_5$label <- "rf5"
explainer_6$label <- "rf6"
explainer_7$label <- "rf7"


fobject <- create_fairness_object(explainer_1,explainer_4,explainer_5,explainer_6,explainer_7,
                                  outcome = "Two_yr_Recidivism",
                                  group  = "Ethnicity",
                                  base   = "Caucasian",
                                  cutoff = 0.5)

fheatmap <- fairness_heatmap(fobject)
plot(fheatmap, flip_axis = TRUE)

}
