% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fairness_object.R
\name{plot_density}
\alias{plot_density}
\title{Plot fairness object}
\usage{
plot_density(x, ...)
}
\arguments{
\item{x}{fairness_object}

\item{...}{other fairness_objects and other parameters}
}
\value{
\code{ggplot} object
}
\description{
Plot distribution for models output probabilities. See how being in particular subgroup affects models decision.
}
\examples{

library(DALEX)
library(ranger)

data("compas")

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
rf_compas2 <- ranger(Two_yr_Recidivism ~Misdemeanor+Age_Above_FourtyFive, data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_rf2 <- explain(rf_compas2, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fobject <-  fairness_check(explainer_glm, explainer_rf2,
                             protected = compas$Ethnicity,
                             privileged = "Caucasian",
                             label = c("lm_1","rf_2"))

plot_density(fobject)
}
