% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fairness_object.R
\name{plot.fairness_object_regression}
\alias{plot.fairness_object_regression}
\title{Plot fairness object regression}
\usage{
\method{plot}{fairness_object_regression}(x, ...)
}
\arguments{
\item{x}{\code{fairness_object_regression} object}

\item{...}{other plot parameters}
}
\value{
\code{ggplot2} object
}
\description{
Plot fairness check enables to look how big differences are between base subgroup (privileged) and unprivileged ones.
If bar plot reaches red zone it means that for this subgroup fairness goal is not satisfied. Multiple subgroups and models can be plotted.
Red and green zone boundary can be moved through epsilon parameter, that needs to be passed through \code{fairness_check}.
}
