% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fairness_pca.R
\name{plot.fairness_pca}
\alias{plot.fairness_pca}
\title{Plot fairness PCA}
\usage{
\method{plot}{fairness_pca}(x, scale = 0.5, ...)
}
\arguments{
\item{x}{fairness_pca object}

\item{scale}{tbd- za mało wiem jak na razie}

\item{...}{other parameters}
}
\value{

}
\description{
Plot fairness PCA
}
