% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fairness_radar.R
\name{plot.fairness_radar}
\alias{plot.fairness_radar}
\title{Plot fairness radar}
\usage{
\method{plot}{fairness_radar}(x, ...)
}
\arguments{
\item{x}{\code{fairness_radar} object}

\item{...}{other \code{fairness_radar} object and other plot parameters}
}
\value{
\code{ggplot} object
}
\description{
Makes radar plot showing different fairness metrics that allow to compare models.
}
\examples{
library(DALEX)
library(ranger)

data("compas")

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fobject <- create_fairness_object(explainer_glm, explainer_rf,
                             outcome = "Two_yr_Recidivism",
                             group = "Ethnicity",
                             base = "Caucasian",
                             cutoff = 0.5)
fradar <- fairness_radar(fobject)
plot(fradar)

}
\references{
code from ModelOriented auditor package, thanks agosiewska!  \url{https://modeloriented.github.io/auditor/}
}
