% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metric_scores.R
\name{plot.metric_scores}
\alias{plot.metric_scores}
\title{Plot metric scores}
\usage{
\method{plot}{metric_scores}(x, ...)
}
\arguments{
\item{x}{\code{metric_scores} object}

\item{...}{other plot parameters}
}
\value{
\code{ggplot} object
}
\description{
Plot metric scores
}
\examples{
data("compas")

# flipping outcomes, 1- favorable
compas$Two_yr_Recidivism <- as.factor(ifelse(compas$Two_yr_Recidivism == '1', '0', '1'))

# train
lm_compas <- glm(Two_yr_Recidivism ~., data = compas, family = binomial())
rf_compas <- ranger::ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)

# numeric target values
y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

# explainer
rf_explainer <- DALEX::explain(rf_compas, data = compas[,-1], y = y_numeric)
lm_explainer <- DALEX::explain(lm_compas, data = compas[,-1], y = y_numeric)

fobject <- fairness_check(rf_explainer, lm_explainer,
                          protected = compas$Ethnicity,
                          privileged = "Caucasian")

ms <- metric_scores(fobject, fairness_metrics = c("TPR","STP","ACC"))
plot(ms)

}
