% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_performance_with_fairness.R
\name{plot_performance_with_fairness}
\alias{plot_performance_with_fairness}
\title{Plot fairness and performance}
\usage{
plot_performance_with_fairness(
  x,
  fairness_metric = NULL,
  performance_metric = NULL
)
}
\arguments{
\item{x}{fairness object}

\item{fairness_metric}{name of fairness metric}

\item{performance_metric}{name of DALEX performance metric}
}
\value{

}
\description{
Plot fairness and performance
}
\examples{

library(DALEX)
library(ranger)

data("compas")

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

rf_compas_1 <- ranger(Two_yr_Recidivism ~Number_of_Priors+Age_Below_TwentyFive, data = compas, probability = TRUE)
lr_compas_1 <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))
rf_compas_2 <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
rf_compas_3 <- ranger(Two_yr_Recidivism ~ Age_Above_FourtyFive+Misdemeanor, data = compas, probability = TRUE)
df <- compas
df$Two_yr_Recidivism <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_1 <- explain(rf_compas_1,  data = compas[,-1], y = y_numeric)
explainer_2 <- explain(lr_compas_1,  data = compas[,-1], y = y_numeric)
explainer_3 <- explain(rf_compas_2,  data = compas[,-1], y = y_numeric, label = "ranger_2")
explainer_4 <- explain(rf_compas_3,  data = compas[,-1], y = y_numeric, label = "ranger_3")

fobject <- create_fairness_object(explainer_1,explainer_2,explainer_3,explainer_4,
                                    data = compas,
                                    outcome = "Two_yr_Recidivism",
                                    group  = "Ethnicity",
                                    base   = "Caucasian")

plot_performance_with_fairness(fobject, "FPR_parity_loss", "auc")

}
