% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stacked_barplot.R
\name{plot_stacked_barplot}
\alias{plot_stacked_barplot}
\alias{plot_stacked_barplot.fairness_object}
\alias{plot_stacked_barplot.deafult}
\title{Title}
\usage{
plot_stacked_barplot(x)

\method{plot_stacked_barplot}{fairness_object}(x)

\method{plot_stacked_barplot}{deafult}(x)
}
\arguments{
\item{x}{fairness object}
}
\value{
ggplot object
}
\description{
Title
}
