% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stacked_barplot.R,
%   R/plot_stacked_metrics.R
\name{plot.stacked_metrics}
\alias{plot.stacked_metrics}
\title{Plot stacked Metrics}
\usage{
\method{plot}{stacked_metrics}(x, ...)

\method{plot}{stacked_metrics}(x, ...)
}
\arguments{
\item{x}{\code{stacked_metrics} object}

\item{...}{other plot parameters}
}
\description{
Plot stacked Metrics
}
