% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_two_models.R
\name{plot_two_models}
\alias{plot_two_models}
\title{Plot two models}
\usage{
plot_two_models(
  fairness_object,
  fairness_metric = NULL,
  performance_metric = NULL
)
}
\arguments{
\item{fairness_object}{object of class \code{fairness_object}}

\item{fairness_metric}{fairness metric name from
\itemize{
\item equal_odds
\item pred_rate_parity
\item acc_parity
\item fnr_parity
\item fpr_parity
\item npv_parity
\item spec_parity
\item mcc_parity
\item model labels
}}

\item{performance_metric}{performance metric name from
\itemize{
\item recall
\item precision
\item accuracy
\item f1
\item auc
}}
}
\value{

}
\description{
Plot two models
}
\examples{
library(DALEX)
library(ranger)
data("compas")
y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

rf_1 <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
lr_1 <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

explainer_1 <- explain(rf_1, data = compas, y = y_numeric)
explainer_2 <- explain(lr_1, data = compas, y = y_numeric)

fo <- create_fairness_object(explainer_1, explainer_2,  outcome = "Two_yr_Recidivism", group = "Ethnicity", base = "Caucasian"  )

plot_two_models(fo, fairness_metric = "fnr_parity", performance_metric = "auc")


}
