% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_ceteris_paribus_cutoff.R
\name{print.ceteris_paribus_cutoff}
\alias{print.ceteris_paribus_cutoff}
\title{Print ceteris paribus cutoff}
\usage{
\method{print}{ceteris_paribus_cutoff}(x, ...)
}
\arguments{
\item{x}{ceteris_paribus_cutoff object}

\item{...}{other print parameters}
}
\value{

}
\description{
Print ceteris paribus cutoff
}
\examples{

library(DALEX)
library(ranger)

data("compas")

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fobject <-create_fairness_object(explainer_glm, explainer_rf,
                                 outcome = "Two_yr_Recidivism",
                                 group = "Ethnicity",
                                 base = "Caucasian",
                                 cutoff = 0.5)

ceteris_paribus_cutoff(fobject, "Asian")
}
