% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_fairness_object.R
\name{print.fairness_object}
\alias{print.fairness_object}
\title{Print Fairness Object}
\usage{
\method{print}{fairness_object}(x, ...)
}
\arguments{
\item{x}{fairness object}

\item{...}{other parameters}
}
\description{
Print Fairness Object
}
