% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_fairness_pca.R
\name{print.fairness_pca}
\alias{print.fairness_pca}
\title{Print fairness PCA}
\usage{
\method{print}{fairness_pca}(x, ...)
}
\arguments{
\item{x}{fairness object}

\item{...}{other params}
}
\value{

}
\description{
Print fairness PCA
}
