% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_fairness_radar.R
\name{print.fairness_radar}
\alias{print.fairness_radar}
\title{Print fairness radar}
\usage{
\method{print}{fairness_radar}(x, ...)
}
\arguments{
\item{x}{fairness_radar object}

\item{...}{other print parameters}
}
\value{

}
\description{
Print fairness radar
}
\examples{

library(DALEX)
library(ranger)

data("compas")

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fobject <-create_fairness_object(explainer_glm, explainer_rf,
                                 outcome = "Two_yr_Recidivism",
                                 group = "Ethnicity",
                                 base = "Caucasian",
                                 cutoff = 0.5)

fairness_radar(fobject)
}
