% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_stacked_metrics.R
\name{print.stacked_metrics}
\alias{print.stacked_metrics}
\title{Print stacked metrics}
\usage{
\method{print}{stacked_metrics}(x, ...)
}
\arguments{
\item{x}{stacked_metrics object}

\item{...}{other stacked_metrics objects}
}
\value{

}
\description{
Stack metrics sums parity loss metrics for all models. Higher value of stacked metrics means the model is less fair (has higher bias)
for subgroups from protected vector.
}
\examples{

data("compas")

# positive outcome - not being recidivist
two_yr_recidivism <- factor(compas$Two_yr_Recidivism, levels = c(1,0))
y_numeric <- as.numeric(two_yr_recidivism) -1

lm_model <- glm(Two_yr_Recidivism~.,
                data=compas,
                family=binomial(link="logit"))

rf_model <- ranger::ranger(Two_yr_Recidivism ~.,
                           data = compas,
                           probability = TRUE,
                           num.trees = 200)

explainer_lm <- DALEX::explain(lm_model, data = compas[,-1], y = y_numeric)
explainer_rf <- DALEX::explain(rf_model, data = compas[,-1], y = y_numeric)

fobject <- fairness_check(explainer_lm, explainer_rf,
                          protected = compas$Ethnicity,
                          privileged = "Caucasian")

sm <- stack_metrics(fobject)
print(sm)
}
