% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_all_cutoffs.R
\name{print.all_cutoffs}
\alias{print.all_cutoffs}
\title{Print all cutoffs}
\usage{
\method{print}{all_cutoffs}(x, ...)
}
\arguments{
\item{x}{all_cuttofs object}

\item{...}{other print parameters}
}
\value{

}
\description{
Print all cutoffs
}
\examples{

data("compas")

# positive outcome - not being recidivist
two_yr_recidivism <- factor(compas$Two_yr_Recidivism, levels = c(1,0))
y_numeric <- as.numeric(two_yr_recidivism) -1

lm_model <- glm(Two_yr_Recidivism~.,
                data=compas,
                family=binomial(link="logit"))

rf_model <- ranger::ranger(Two_yr_Recidivism ~.,
                           data = compas,
                           probability = TRUE,
                           num.trees = 200)

explainer_lm <- DALEX::explain(lm_model, data = compas[,-1], y = y_numeric)
explainer_rf <- DALEX::explain(rf_model, data = compas[,-1], y = y_numeric)

fobject <- fairness_check(explainer_lm, explainer_rf,
                          protected = compas$Ethnicity,
                          privileged = "Caucasian")

ac <- all_cutoffs(fobject,
                  label = "lm",
                  fairness_metrics = c("TPR_parity_loss",
                                       "FPR_parity_loss"))
print(ac)

}
