% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_chosen_metric.R
\name{print.chosen_metric}
\alias{print.chosen_metric}
\title{Print chosen metric}
\usage{
\method{print}{chosen_metric}(x, ...)
}
\arguments{
\item{x}{\code{chosen_metric} object}

\item{...}{other \code{chosen_metric} object}
}
\value{

}
\description{
Print chosen metric
}
\examples{

library(DALEX)
library(ranger)

data(compas)

rf_compas <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE) # Wszystko
lr_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))
# numeric target values
y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

# explainer
rf_explainer <- explain(rf_compas, data = compas[,-1], y = y_numeric)
lr_explainer <- explain(lr_compas, data = compas[,-1], y = y_numeric)


fobject <- create_fairness_object(rf_explainer, lr_explainer,
                                  data = compas,
                                  outcome = "Two_yr_Recidivism",
                                  group = "Ethnicity",
                                  base = "African_American")

cm <- choose_metric(fobject, "TPR_parity_loss")
print(cm)
}
