% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_fairness_check.R
\name{print.fairness_check}
\alias{print.fairness_check}
\title{Print fairness check}
\usage{
\method{print}{fairness_check}(x, ...)
}
\arguments{
\item{x}{fairness_check object}

\item{...}{other plot parameters}
}
\value{

}
\description{
Print fairness check
}
\examples{

library(DALEX)
library(ranger)

data("compas")

# chaging levels - positive situation here is not becoming recidivist
compas$Two_yr_Recidivism <- factor(compas$Two_yr_Recidivism, levels = c(1,0))

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fobject <-create_fairness_object(explainer_glm, explainer_rf,
                                 outcome = "Two_yr_Recidivism",
                                 data  = compas,
                                 group = "Sex",
                                 base  = "Female",
                                cutoff = 0.5)

fairness_check(fobject)

fobject2 <-create_fairness_object(explainer_glm, explainer_rf,
                                 outcome = "Two_yr_Recidivism",
                                 data  = compas,
                                 group = "Sex",
                                 base  = "Female",
                                 cutoff = c(0.5,0.6) )
fairness_check(fobject2)

}
