% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_group_metrics.R
\name{print.group_metric}
\alias{print.group_metric}
\title{Print group metric}
\usage{
\method{print}{group_metric}(x, ...)
}
\arguments{
\item{x}{\code{group_metric} object}

\item{...}{other print parameters}
}
\description{
Print group metric
}
\examples{

library(DALEX)
library(ranger)
data("compas")
y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

rf_1 <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
lr_1 <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

explainer_1 <- explain(rf_1, data = compas, y = y_numeric)
explainer_2 <- explain(lr_1, data = compas, y = y_numeric)

fo <- create_fairness_object(explainer_1, explainer_2,  outcome = "Two_yr_Recidivism", group = "Ethnicity", base = "Caucasian"  )

group_metric(fo, fairness_metric = "FPR", performance_metric = "auc")
}
