% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.metric_scores}
\alias{print.metric_scores}
\title{Print metric scores data}
\usage{
\method{print}{metric_scores}(x, ...)
}
\arguments{
\item{x}{\code{metric_scores} object}

\item{...}{other print parameters}
}
\description{
Print metric scores data
}
\examples{

data("compas")

# flipping outcomes, 1- favorable
compas$Two_yr_Recidivism <- as.factor(ifelse(compas$Two_yr_Recidivism == '1', '0', '1'))

# train
lm_compas <- glm(Two_yr_Recidivism ~., data = compas, family = binomial())
rf_compas <- ranger::ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)

# numeric target values
y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

# explainer
rf_explainer <- DALEX::explain(rf_compas, data = compas[,-1], y = y_numeric)
lm_explainer <- DALEX::explain(lm_compas, data = compas[,-1], y = y_numeric)

fobject <- fairness_check(rf_explainer, lm_explainer,
                          protected = compas$Ethnicity,
                          privileged = "Caucasian")

ms <- metric_scores(fobject, fairness_metrics = c("TPR","STP","ACC"))
ms
}
