% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_two_models.R
\name{print.two_models}
\alias{print.two_models}
\title{Print Two models object}
\usage{
\method{print}{two_models}(x, ...)
}
\arguments{
\item{x}{two_models object}

\item{...}{other parameters}
}
\value{

}
\description{
Print Two models object
}
\examples{

library(DALEX)
library(ranger)

data("compas")

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

rf_1 <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
lr_1 <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

explainer_1 <- explain(rf_1, data = compas, y = y_numeric)
explainer_2 <- explain(lr_1, data = compas, y = y_numeric)

two_m <- two_models(explainer_1, explainer_2,
                    fairness_metric = "fpr_parity",
                    outcome = "Two_yr_Recidivism",
                    group  = "Ethnicity",
                    base   = "Caucasian")

print(two_m)
}
