% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{regresion_metrics}
\alias{regresion_metrics}
\title{Regression metrics}
\usage{
regresion_metrics(explainer, protected, privileged)
}
\arguments{
\item{privileged}{}
}
\value{

}
\description{
Regression metrics
}
\examples{

}
