% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_metrics.R
\name{regression_metrics}
\alias{regression_metrics}
\title{Regression metrics}
\usage{
regression_metrics(explainer, protected, privileged)
}
\arguments{
\item{explainer}{object created with \code{\link[DALEX]{explain}}}

\item{protected}{factor, protected variable (also called sensitive attribute), containing privileged and unprivileged groups}

\item{privileged}{factor/character, one value of \code{protected}, denoting subgroup suspected of the most privilege}
}
\value{
\code{data.frame}
}
\description{
Regression metrics
}
