% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Resample}
\usage{
resample(protected, y, type = "uniform", probs = NULL)
}
\arguments{
\item{protected}{factor, protected variables with subgroups as levels (sensitive attributes)}

\item{y}{numeric, vector with classes 0 and 1, where 1 means favorable class.}

\item{type}{character, either (default) 'uniform' or 'preferential'}

\item{probs}{numeric, vector with probabilities for preferential sampling}
}
\value{
numeric vector of indexes
}
\description{
Method of bias mitigation. Similarly to \code{reweight_classes} this method computes desired number of observations if the protected variable is independent
from y and on this basis decides if this subgroup with certain class (+ or -) should be more or less numerous. Than performs oversampling or undersampling depending on the case.
}
\references{
This method was implemented based on Kamiran, Calders 2011 \url{https://link.springer.com/content/pdf/10.1007/s10115-011-0463-8.pdf}
}
