% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reweight_classes.R
\name{reweight_classes}
\alias{reweight_classes}
\title{Reweight classes}
\usage{
reweight_classes(protected, y)
}
\arguments{
\item{protected}{factor, protected variables with subgroups as levels (sensitive attributes)}

\item{y}{numeric, vector with classes 0 and 1, where 1 means favorable class.}
}
\value{
numeric, vector of weights
}
\description{
Function returns weights for model training. The purpose of this weights is to mitigate bias in statistical parity w
In fact this could worsen the performance in other fairness metrics. This affects also model's performance metrics (accuracy).
}
\details{
Method produces weights for each subgroup for each class. Firstly assumes that protected variable and class are independent and
calculates expected probability of this certain event (that subgroup == a and class = c). Than it calculates the actual probability of this event based
on empirical data. Finally the weight is quotient of those probabilities
}
\references{
This method was implemented based on Kamiran, Calders 2011 \url{https://link.springer.com/content/pdf/10.1007/s10115-011-0463-8.pdf}
}
