% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_pivot.R
\name{roc_pivot}
\alias{roc_pivot}
\title{Reject Option based Classification pivot}
\usage{
roc_pivot(x, protected, privileged, cutoff = 0.5, theta = 0.1)
}
\arguments{
\item{x}{DALEX explainer}

\item{protected}{factor, protected variables with subgroups as levels (sensitive attributes)}

\item{privileged}{factor/character, level in protected denoting privileged subgroup}

\item{cutoff}{numeric, threshold for all subgroups}

\item{theta}{numeric, variable specifies maximal euclidean distance to cutoff resulting ing label switch}
}
\value{
DALEX explainer (with changed y_hat)
}
\description{
Reject Option based Classifier is postprocessing bias mitigation method. Method changes labels of favourable, privileged and close to cutoff observations to unfavourable
and the opposite for unprivileged observations (changing unfavourable and close to cutoff observations to favourable, more in details).
By this potentially wrongfully labbeled observations are assigned different labels.
Note that in y in DALEX explainer 1 should indicate favourable outcome.
}
\details{
Method implemented implemented based on article (Kamiran, Karim, Zhang 2012). In original implementation labels should be switched. Due to specific DALEX methods
probabilities (y_hat) are assigned value in equal distance but other side of cutoff. The method changes explainers y_hat values in two cases.
1. When unprivileged subgroup is within (cutoff - theta, cutoff)
2. When privileged subgroup is within (cutoff, cutoff + theta)
}
\references{
Kamiran, Karim, Zhang 2012 \url{https://ieeexplore.ieee.org/document/6413831/} ROC method
}
