% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_models.R
\name{two_models}
\alias{two_models}
\title{Two models plot}
\usage{
two_models(
  x,
  y,
  outcome,
  group,
  base,
  data = NULL,
  fairness_metric = NULL,
  performance_metric = NULL,
  cutoff = 0.5
)
}
\arguments{
\item{x}{explainer 1}

\item{y}{explainer 2}

\item{outcome}{target variable}

\item{group}{protected group/ variable}

\item{base}{threshold to which metrics will be compared}

\item{data}{data frame, if null getting data from \code{x}}

\item{fairness_metric}{fairness metric name}

\item{performance_metric}{performance metric name}

\item{cutoff}{deafult = 0.5}
}
\value{

}
\description{
Two models plot
}
