% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_cutoffs.R
\name{all_cutoffs}
\alias{all_cutoffs}
\title{All cutoffs}
\usage{
all_cutoffs(
  x,
  grid_points = 101,
  fairness_metrics = unique_metrics(),
  explainer_label
)
}
\arguments{
\item{x}{fairness_object}

\item{grid_points}{numeric, grid for cutoffs to test. Number of points between 0 and 1 spread evenly.}

\item{fairness_metrics}{character, name of metric or vector of metrics}

\item{explainer_label}{character, single label of choosen explainer}
}
\value{
all_cutoffs object
}
\description{
Create all_cutoffs object and to show how with change of cutoffs parity metrics change for single model/explainer.
It is highly suggested to pick small number of metrics. With value of cutoff all subgroup cutoffs change.
}
\examples{

library(DALEX)
library(ranger)

data("compas")

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fobject <-create_fairness_object(explainer_glm, explainer_rf,
                                 outcome = "Two_yr_Recidivism",
                                 group = "Ethnicity",
                                 base = "Caucasian",
                                 cutoff = 0.5)

ac <- all_cutoffs(fobject, fairness_metrics = c("TPR_parity_loss", "F1_parity_loss"), explainer_label = "ranger")
plot(ac)

}
