% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_group_fairness_metrics.R
\name{calculate_group_fairness_metrics}
\alias{calculate_group_fairness_metrics}
\title{Calculate fairness metrics in groups}
\usage{
calculate_group_fairness_metrics(x)
}
\arguments{
\item{x}{object of class \code{group_matrices}}
}
\value{
\code{group_metric_matrix} object
It's a \code{data.frame} with metrics as row names and scores for those metrics for each subgroup in columns
}
\description{
Create \code{data.frame} from \code{group_matrices} object containing metric scores for each subgroup.
}
