% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Confusion matrix}
\usage{
confusion_matrix(predicted, observed, true_value)
}
\arguments{
\item{predicted}{predicted from probs}

\item{observed}{actual values}

\item{true_value}{true}
}
\description{
Confusion matrix
}
