% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_matrix.R
\name{confusion_matrix}
\alias{confusion_matrix}
\title{Confusion matrix}
\usage{
confusion_matrix(probs, observed, cutoff = 0.5)
}
\arguments{
\item{probs}{probabilities given by model}

\item{observed}{actual values from outcome}

\item{cutoff}{cutoff}
}
\description{
Confusion matrix
}
