% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fairness_object.R
\name{create_fairness_object}
\alias{create_fairness_object}
\title{Create Fairness Object}
\usage{
create_fairness_object(
  x,
  ...,
  data = NULL,
  outcome,
  group,
  base = NULL,
  cutoff = 0.5
)
}
\arguments{
\item{x}{DALEX explainer}

\item{...}{more dalex explainers}

\item{data}{Data frame to be used}

\item{outcome}{the target of classification}

\item{group}{protected group/variable that}

\item{base}{in regard to what subgroup of group}

\item{cutoff}{threshold for probability, deafult 0.5}
}
\value{
An object of class \code{fairness object}

It's a list with following fields:
}
\description{
Create Fairness Object
}
\examples{
library(DALEX)
library(ranger)

data("compas")

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

cfo <-create_fairness_object(explainer_glm, explainer_rf,
                             outcome = "Two_yr_Recidivism",
                             group = "Ethnicity",
                             base = "Caucasian",
                             cutoff = 0.5)


}
