% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fairness_object.R
\name{create_fairness_object}
\alias{create_fairness_object}
\title{Create Fairness Object}
\usage{
create_fairness_object(
  x,
  ...,
  data = NULL,
  outcome,
  group,
  base = NULL,
  cutoff = NULL
)
}
\arguments{
\item{x}{DALEX explainer}

\item{...}{possibly more DALEX explainers}

\item{data}{Data frame to be used}

\item{outcome}{the target of classification}

\item{group}{protected group/variable that}

\item{base}{in regard to what subgroup parity loss is calculated}

\item{cutoff}{threshold for probability, can be vector of thresholds with diferent value for each level of group, deafult 0.5}
}
\value{
An object of class \code{fairness object} which is a list with elements:

\itemize{
\item metric_data data frame created with following metrics:
\itemize{

\item TPR - True Positive Rate (Sensitivity, Recall, Equal Odds)
\item TNR - True Negative Rate (Specificity)
\item PPV - Positive Predictive Value (Precision)
\item NPV - Negative Predictive Value
\item FNR - False Negative Rate
\item FPR - False Positive Rate
\item FDR - False Discovery Rate
\item FOR - False Omision Rate
\item TS - Threat Score
\item ACC - Accuracy
\item F1 - F1 Score
\item MCC - Matthews correlation coefficient
}

M_parity_loss = sum(abs(metric - base_metric))

where:

M - some metric mentioned above

metric - vector of metrics from each subgroup

base_metric - scalar, value of metric for base subgroup

\item groups_data - metrics across groups, with base but not summarised for explainer

\item explainers  - list of DALEX explainers

\item data - data

\item ...         - other parameters passed to function
}
}
\description{
Create fairness object which measures different fairness metrics and wraps data, explainers and parameters in useful object.
}
\examples{
library(DALEX)
library(ranger)

data("compas")

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

cfo <-create_fairness_object(explainer_glm, explainer_rf,
                             outcome = "Two_yr_Recidivism",
                             group = "Ethnicity",
                             base = "Caucasian",
                             cutoff = 0.5)


}
\references{
Fairness object took inspiration from R package \code{fairness} by kozodoi. Some parameters where designed to show resemblance.
}
