% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_parity.R
\name{demographic_parity}
\alias{demographic_parity}
\alias{demographic_parity.default}
\alias{demographic_parity.fairness_object}
\title{Demographic parity}
\usage{
demographic_parity(x, ...)

\method{demographic_parity}{default}(x, ..., outcome, group)

\method{demographic_parity}{fairness_object}(x, ...)
}
\arguments{
\item{x}{data.frame/ fairness_object}

\item{...}{other parameters}

\item{outcome}{target column name in data}

\item{group}{protected group (column) in data}
}
\value{

}
\description{
Demographic parity
}
\examples{

data("compas")
demographic_parity(compas, outcome =  "Two_yr_Recidivism", group = "Ethnicity")

}
