% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disparate_impact_remover.R
\name{disparate_impact_remover}
\alias{disparate_impact_remover}
\title{Disparate impact remover}
\usage{
disparate_impact_remover(data, protected, features_to_transform, lambda)
}
\arguments{
\item{data}{data.frame, data to be transformed}

\item{protected}{factor, vector containing sensitive information}

\item{features_to_transform}{character, vector of column names to be transformed. Columns must have numerical, ordinal values}

\item{lambda}{numeric, amount of repair desired. Value from 0 to 1, where 0 will return unchanged dataset and 1 fully repaired dataset}
}
\value{
data.frame
}
\description{
Disparate impact remover is a method of bias mitigation. It changes distribution ordinal features of data with regard to earth mover distance.
}
\references{
https://arxiv.org/pdf/1412.3756.pdf
}
