% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_fairness_object.R,
%   R/expand_fobject_data.R
\name{expand_fairness_object}
\alias{expand_fairness_object}
\title{Expand fairness object}
\usage{
expand_fairness_object(x)

expand_fairness_object(x)
}
\arguments{
\item{x}{fairness_object}
}
\value{
data frame more feasible to plot


}
\description{
Expand fairness object

Expand Fairness Object
}
