% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_fairness_object.R
\name{expand_fairness_object}
\alias{expand_fairness_object}
\title{Expand Fairness Object}
\usage{
expand_fairness_object(x, scale = FALSE, drop_metrics_with_na = FALSE)
}
\arguments{
\item{x}{fairness_object}

\item{scale}{logical, if \code{TRUE} standarised.}

\item{drop_metrics_with_na}{logical, if TRUE metrics with NA will be omitted}
}
\value{

}
\description{
Unfold fairness object to 3 columns (metrics, label, score) to construct better base for visualization.
}
\examples{

library(DALEX)
library(ranger)

data("compas")

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fobject <-create_fairness_object(explainer_glm, explainer_rf,
                             outcome = "Two_yr_Recidivism",
                             group = "Ethnicity",
                             base = "Caucasian",
                             cutoff = 0.5)

expand_fairness_object(fobject, drop_metrics_with_na = TRUE)

}
