% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_fairness_object.R
\name{expand_fairness_object}
\alias{expand_fairness_object}
\title{Expand Fairness Object}
\usage{
expand_fairness_object(x, scale = FALSE)
}
\arguments{
\item{x}{fairness_object}

\item{scale}{if true normalises metric data}
}
\value{

}
\description{
Expand Fairness Object
}
