% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairness_check.R
\name{fairness_check}
\alias{fairness_check}
\title{Fairness check}
\usage{
fairness_check(x, epsilon = 0.1)
}
\arguments{
\item{x}{fairness_object}

\item{epsilon}{numeric, margins for decision if fair or not. Default 0.1}
}
\value{

}
\description{
Metrics used are made for each subgroup, then base metric score is subtracted leaving loss of particular metric.
If loss is less than -epsilon than such metric is marked as "not passed". There is no upper boundary for fairness, so if some subgroup is "better"
in particular metric than privileged one there are no consequences to that. Good metric description can be found here :
If loss is greater than absolute value of epsilon than such metric is marked as "not passed". It means that values of metrics should be within (-epsilon,epsilon) boundary.
Epsilon value can be adjusted to user's needs. There are some metrics that might be derived from existing metrics (For example Equalized Odds - equal TPR and FPR for all subgroups).
That means passing 5 metrics in fairness check asserts that model is even more fair. In \code{fairness_check} models must always predict commonly desired result. Not adhering to this rule
may lead to misinterpretation of the plot. More on metrics and their equivalents:
\url{https://fairware.cs.umass.edu/papers/Verma.pdf}
\url{https://en.wikipedia.org/wiki/Fairness_(machine_learning)}
}
\examples{

library(DALEX)
library(ranger)

data("compas")

# changing levels - positive situation here is not becoming recidivist
compas$Two_yr_Recidivism <- factor(compas$Two_yr_Recidivism, levels = c(1,0))

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fobject <-create_fairness_object(explainer_glm, explainer_rf,
                                 outcome = "Two_yr_Recidivism",
                                 data  = compas,
                                 group = "Sex",
                                 base  = "Female")

# changing cutoff values
fobject2 <-create_fairness_object(explainer_glm, explainer_rf,
                                 outcome = "Two_yr_Recidivism",
                                 data  = compas,
                                 group = "Sex",
                                 base  = "Female",
                                cutoff = c(0.45,0.5),
                                label = c("lm_c","rf_c")) # c for changed

fc <- fairness_check(fobject)
fc2 <- fairness_check(fobject2)

plot(fc,fc2)

}
\references{
Zafar,Valera, Rodriguez, Gummadi (2017)  \url{https://arxiv.org/pdf/1610.08452.pdf}

Hardt, Price, Srebro (2016) \url{https://arxiv.org/pdf/1610.02413.pdf}

Verma, Rubin (2018) \url{https://fairware.cs.umass.edu/papers/Verma.pdf}
}
