% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairness_check.R
\name{fairness_check}
\alias{fairness_check}
\title{Fairness check}
\usage{
fairness_check(
  x,
  ...,
  protected,
  privileged,
  cutoff = NULL,
  label = NULL,
  epsilon = NULL
)
}
\arguments{
\item{x}{\code{DALEX explainer}/\code{fairness_object}}

\item{...}{possibly more \code{DALEX explainers}/\code{fairness_objects}}

\item{protected}{vector, protected variable (also called sensitive attribute), containing privileged and unprivileged groups}

\item{privileged}{one value of \code{protected}, in regard to what subgroup parity loss is calculated}

\item{cutoff}{numeric, vector of cutoffs (thresholds) for each value of protected variable, affecting only explainers.}

\item{label}{character, vector of labels to be assigned for explainers, default is explainer label.}
}
\value{
An object of class \code{fairness_object} which is a list with elements:
\itemize{
\item metric_data - data.frame containing parity loss for various fairness metrics. Created with following metrics:
\itemize{

\item TPR - True Positive Rate (Sensitivity, Recall, Equal Odds)
\item TNR - True Negative Rate (Specificity)
\item PPV - Positive Predictive Value (Precision)
\item NPV - Negative Predictive Value
\item FNR - False Negative Rate
\item FPR - False Positive Rate
\item FDR - False Discovery Rate
\item FOR - False Omission Rate
\item TS  - Threat Score
\item STP - Statistical Parity
\item ACC - Accuracy
\item F1  - F1 Score
\item MCC - Matthews correlation coefficient
}

M_parity_loss = sum(abs(metric - base_metric))

where:

M - some metric mentioned above

metric - vector of metrics from each subgroup

base_metric - scalar, value of metric for base subgroup

\item groups_data - metrics across levels in protected variable

\item explainers  - list of DALEX explainers used to create object

\item ...         - other parameters passed to function
}
}
\description{
Fairness check creates fairness object which measures different fairness metrics and wraps data, explainers and parameters in useful object. This is fundamental object in this package
It allows to visualize fairness metrics in many ways and compare models on both fairness and performance level. Fairness check acts as merger and wrapper for explainers and fairness objects.
While other fairness objects values are not changed, fairness check assigns cutoffs and labels to provided explainers so same explainers with changed labels/cutoffs might be gradually added to fairness object.
Users through print and plot methods may quickly check values of most popular fairness metrics. More on that topic in details/
}
\details{
Metrics used are made for each subgroup, then base metric score is subtracted leaving loss of particular metric.
If absolute loss is greater than epsilon than such metric is marked as "not passed". It means that values of metrics should be within (-epsilon,epsilon) boundary.
Epsilon value can be adjusted to user's needs. There are some metrics that might be derived from existing metrics (For example Equalized Odds - equal TPR and FPR for all subgroups).
That means passing 5 metrics in fairness check asserts that model is even more fair. In \code{fairness_check} models must always predict positive result. Not adhering to this rule
may lead to misinterpretation of the plot. More on metrics and their equivalents:
\url{https://fairware.cs.umass.edu/papers/Verma.pdf}
\url{https://en.wikipedia.org/wiki/Fairness_(machine_learning)}
}
\examples{
library(DALEX)
library(ranger)

data("compas")

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas,
 family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fcheck <- fairness_check(explainer_glm, explainer_rf,
                             protected = compas$Ethnicity,
                             privileged = "Caucasian")
plot(fcheck)

}
\references{
Zafar,Valera, Rodriguez, Gummadi (2017)  \url{https://arxiv.org/pdf/1610.08452.pdf}

Hardt, Price, Srebro (2016) \url{https://arxiv.org/pdf/1610.02413.pdf}

Verma, Rubin (2018) \url{https://fairware.cs.umass.edu/papers/Verma.pdf}
}
