% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_matrices.R
\name{group_matrices}
\alias{group_matrices}
\title{Group confusion matrices}
\usage{
group_matrices(protected, probs, preds, cutoff)
}
\arguments{
\item{probs}{\code{character} name of column with probabilities}

\item{cutoff}{\code{numeric} cutoff for probabilities, default = 0.5}

\item{data}{data frame}

\item{group}{\code{character} name of column with group}

\item{outcome}{\code{character} name of column with outcome}

\item{outcome_numeric}{\code{numeric} vector of outcome}
}
\value{
list with values:
\itemize{
\item subgroup
\itemize{
\item tp - true positive values
\item fp - false positive values
\item tn - true negative values
\item fn - false negative values
}}
}
\description{
Calculates confusion matrices for each subgroup
}
\examples{
data("compas")

glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))
y_prob <- glm_compas$fitted.values

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

gm <- group_matrices(as.numeric(compas$Ethnicity),
                     y_prob,
                     y_numeric,
                     cutoff = rep(0.45,6))

gm

}
