% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_matrices.R
\name{group_matrices}
\alias{group_matrices}
\title{Group confusion matrices}
\usage{
group_matrices(
  data,
  group,
  probs = "probabilities",
  true_value = 1,
  outcome,
  outcome_numeric,
  cutoff
)
}
\arguments{
\item{data}{data frame}

\item{group}{\code{character} name of column with group}

\item{probs}{\code{character} name of column with probabilities}

\item{true_value}{value to be considered as positive case}

\item{outcome}{\code{character} name of column with outcome}

\item{outcome_numeric}{\code{numeric} vector of outcome}

\item{cutoff}{\code{numeric} cutoff for probabilities, deafult = 0.5}
}
\value{

}
\description{
Group confusion matrices
}
