% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_model_performance.R
\name{group_model_performance}
\alias{group_model_performance}
\title{Group model performance}
\usage{
group_model_performance(x, protected, cutoff, performance_metric)
}
\arguments{
\item{x}{object created with \code{\link[DALEX]{explain}}}

\item{protected}{factor, vector with levels as subgroups}

\item{cutoff}{vector of thresholds for each subgroup}

\item{performance_metric}{name of performance metric}
}
\value{
score in performance metric between 0 and 1
}
\description{
Special method for model performance evaluation. Counts number of tp, tn, fp, fn for each subgroup (and therefore potentially distinct cutoff), sums afterwards.
}
