% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_cutoffs.R
\name{plot.all_cutoffs}
\alias{plot.all_cutoffs}
\title{Plot all cutoffs}
\usage{
\method{plot}{all_cutoffs}(x, ...)
}
\arguments{
\item{x}{all_cutoffs object}

\item{...}{other plot parameters}
}
\value{
ggplot object
}
\description{
All cutoffs plot allows to check how parity loss of chosen metrics is affected by the change of cutoff. Values of cutoff
are the same for all subgroups (levels of protected variable) no matter what cuttof values were in fairness_object.
}
\examples{

data("german")

y_numeric <- as.numeric(german$Risk) -1

lm_model <- glm(Risk~.,
                data = german,
                family=binomial(link="logit"))

rf_model <- ranger::ranger(Risk ~.,
                           data = german,
                           probability = TRUE,
                           num.trees = 200)

explainer_lm <- DALEX::explain(lm_model, data = german[,-1], y = y_numeric)
explainer_rf <- DALEX::explain(rf_model, data = german[,-1], y = y_numeric)

fobject <- fairness_check(explainer_lm, explainer_rf,
                          protected = german$Sex,
                          privileged = "male")

ac <- all_cutoffs(fobject,
                  label = "lm",
                  fairness_metrics = c("TPR_parity_loss",
                                       "FPR_parity_loss"))
plot(ac)


}
