% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_all_cutoffs.R
\name{plot.all_cutoffs}
\alias{plot.all_cutoffs}
\title{Plot all cutoffs}
\usage{
\method{plot}{all_cutoffs}(x, ...)
}
\arguments{
\item{x}{all_cutoffs object}

\item{...}{other plot parameters}
}
\value{
ggplot object
}
\description{
Plot all cutoffs
}
\examples{

library(DALEX)
library(ranger)

data("compas")

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fobject <-create_fairness_object(explainer_glm, explainer_rf,
                                 outcome = "Two_yr_Recidivism",
                                 group = "Ethnicity",
                                 base = "Caucasian",
                                 cutoff = 0.5)

ac <- all_cutoffs(fobject, fairness_metrics = c("TPR_parity_loss", "F1_parity_loss"), explainer_label = "ranger")
plot(ac)


}
