% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ceteris_paribus_cutoff.R
\name{plot.ceteris_paribus_cutoff}
\alias{plot.ceteris_paribus_cutoff}
\title{Ceteris paribus cutoff plot}
\usage{
\method{plot}{ceteris_paribus_cutoff}(x, ...)
}
\arguments{
\item{x}{ceteris_paribus_cutoff object}

\item{...}{other ceteris_paribus_cutoff objects}
}
\value{
ggplot object
}
\description{
Ceteris paribus cutoff plot
}
\examples{

library(DALEX)
library(ranger)

data("compas")

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fobject <-create_fairness_object(explainer_glm, explainer_rf,
                                 outcome = "Two_yr_Recidivism",
                                 group = "Ethnicity",
                                 base = "Caucasian",
                                 cutoff = 0.5)

cpc <- ceteris_paribus_cutoff(fobject, subgroup = "African_American")
plot(cpc)


cpc <- ceteris_paribus_cutoff(fobject, subgroup = "African_American",
                              cummulated = TRUE,
                              fairness_metrics = c("TPR_parity_loss","PPV_parity_loss", "TNR_parity_loss" ))
plot(cpc)


}
