% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_metrics.R
\name{stack_metrics}
\alias{stack_metrics}
\title{Stack metrics}
\usage{
stack_metrics(x)
}
\arguments{
\item{x}{\code{fairness_object}}
}
\value{
\code{stacked_metrics} object
}
\description{
Stack metrics sums parity loss metrics for all models. Higher value of stacked metrics means the model is less fair (has higher bias)
for subgroups from protected vector.
}
\examples{

data("compas")

# positive outcome - not being recidivist
two_yr_recidivism <- factor(compas$Two_yr_Recidivism, levels = c(1,0))
y_numeric <- as.numeric(two_yr_recidivism) -1

lm_model <- glm(Two_yr_Recidivism~.,
                data=compas,
                family=binomial(link="logit"))

rf_model <- ranger::ranger(Two_yr_Recidivism ~.,
                           data = compas,
                           probability = TRUE,
                           num.trees = 200)

explainer_lm <- DALEX::explain(lm_model, data = compas[,-1], y = y_numeric)
explainer_rf <- DALEX::explain(rf_model, data = compas[,-1], y = y_numeric)

fobject <- fairness_check(explainer_lm, explainer_rf,
                          protected = compas$Ethnicity,
                          privileged = "Caucasian")

sm <- stack_metrics(fobject)
plot(sm)
}
