% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stacked_barplot.R
\name{plot_stacked_barplot}
\alias{plot_stacked_barplot}
\alias{plot_stacked_barplot.fairness_object}
\alias{plot_stacked_barplot.deafult}
\title{Title}
\usage{
plot_stacked_barplot(x)

\method{plot_stacked_barplot}{fairness_object}(x)

\method{plot_stacked_barplot}{deafult}(x)
}
\arguments{
\item{x}{fairness object}
}
\value{
ggplot object
}
\description{
Title
}
\examples{

library(DALEX)
library(ranger)

data("compas")

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fobject <-create_fairness_object(explainer_glm, explainer_rf,
                             outcome = "Two_yr_Recidivism",
                             group = "Ethnicity",
                             base = "Caucasian",
                             cutoff = 0.5)

plot_stacked_barplot(fobject)

}
