% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack_metrics.R
\name{stack_metrics}
\alias{stack_metrics}
\title{Stack metrics}
\usage{
stack_metrics(x)
}
\arguments{
\item{x}{\code{fairness_object}}
}
\value{
\code{stacked_metrics} object
}
\description{
Stacked metrics is like plot for \code{choosen_metric} but with all unique metrics stacked on top of each other. Metrics containing NA's will be dropped to enable fair comparison.
}
\examples{

library(DALEX)
library(ranger)

data("compas")

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fobject <-create_fairness_object(explainer_glm, explainer_rf,
                             outcome = "Two_yr_Recidivism",
                             group = "Ethnicity",
                             base = "Caucasian",
                             cutoff = 0.5)

sm <- stack_metrics(fobject)
plot(sm)
}
