% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stacked_metrics.R
\name{plot.stacked_metrics}
\alias{plot.stacked_metrics}
\title{Plot stacked Metrics}
\usage{
\method{plot}{stacked_metrics}(x, ...)
}
\arguments{
\item{x}{\code{stacked_metrics} object}

\item{...}{other \code{stacked_metrics} objects and other plot parameters}
}
\description{
Plot stacked Metrics
}
\examples{

library(DALEX)
library(ranger)

data("compas")

rf_compas  <- ranger(Two_yr_Recidivism ~., data = compas, probability = TRUE)
glm_compas <- glm(Two_yr_Recidivism~., data=compas, family=binomial(link="logit"))

y_numeric <- as.numeric(compas$Two_yr_Recidivism)-1

explainer_rf  <- explain(rf_compas, data = compas, y = y_numeric)
explainer_glm <- explain(glm_compas, data = compas, y = y_numeric)

fobject <-create_fairness_object(explainer_glm, explainer_rf,
                             outcome = "Two_yr_Recidivism",
                             group = "Ethnicity",
                             base = "Caucasian",
                             cutoff = 0.5)

sm <- stack_metrics(fobject)
plot(sm)


}
