% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stacked_metrics.R
\name{plot.stacked_metrics}
\alias{plot.stacked_metrics}
\title{Plot stacked Metrics}
\usage{
\method{plot}{stacked_metrics}(x, ...)
}
\arguments{
\item{x}{\code{stacked_metrics} object}

\item{...}{other plot parameters}
}
\description{
Stacked metrics is like plot for \code{choosen_metric} but with all unique metrics stacked on top of each other.
Metrics containing NA's will be dropped to enable fair comparison.
}
\examples{

data("compas")

# positive outcome - not being recidivist
two_yr_recidivism <- factor(compas$Two_yr_Recidivism, levels = c(1,0))
y_numeric <- as.numeric(two_yr_recidivism) -1

lm_model <- glm(Two_yr_Recidivism~.,
                data=compas,
                family=binomial(link="logit"))

rf_model <- ranger::ranger(Two_yr_Recidivism ~.,
                           data = compas,
                           probability = TRUE,
                           num.trees = 200)

explainer_lm <- DALEX::explain(lm_model, data = compas[,-1], y = y_numeric)
explainer_rf <- DALEX::explain(rf_model, data = compas[,-1], y = y_numeric)

fobject <- fairness_check(explainer_lm, explainer_rf,
                          protected = compas$Ethnicity,
                          privileged = "Caucasian")

sm <- stack_metrics(fobject)
plot(sm)

}
