% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_two_models.R
\name{plot.two_models}
\alias{plot.two_models}
\title{Plot Two Models}
\usage{
\method{plot}{two_models}(x, ...)
}
\arguments{
\item{x}{two_models object}

\item{...}{other parameters}
}
\value{
ggplot object
}
\description{
Plot Two Models
}
